common: rename utils-xiph crate to common, see comments PR #318
