+++ symphonia-bundle-flac/Cargo.toml
@@ -16,7 +16,7 @@ rust-version = "1.56"
 log = "0.4"
 symphonia-core = { version = "0.5.4", path = "../symphonia-core" }
 symphonia-metadata = { version = "0.5.4", path = "../symphonia-metadata" }
-symphonia-utils-xiph = { version = "0.5.4", path = "../symphonia-utils-xiph" }
+symphonia-common = { version = "0.5.4", path = "../symphonia-common" }
 
 [lints.rust]
 unexpected_cfgs = { level = "warn", check-cfg = ['cfg(fuzzing)'] }
