+++ symphonia-format-mkv/Cargo.toml
@@ -17,4 +17,4 @@ log = "0.4"
 lazy_static = "1.4.0"
 symphonia-core = { version = "0.5.4", path = "../symphonia-core" }
 symphonia-metadata = { version = "0.5.4", path = "../symphonia-metadata" }
-symphonia-utils-xiph = { version = "0.5.4", path = "../symphonia-utils-xiph" }
+symphonia-common = { version = "0.5.4", path = "../symphonia-common" }
